change: amountカラムをint型に変更 - 数値計算が複雑でないため

+++ database/migrations/2025_12_09_013303_create_incomes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                    ->unique()
                    ->constrained()
                    ->cascadeOnDelete();
-            $table->decimal('amount', 12, 2);
+            $table->integer('amount');
             $table->timestamps();
         });
     }
